requirements.txt
fastapi
uvicorn[standard]
pandas
sqlalchemy
smartapi-python
logzero
pyotp
smartmoneyconcepts
ta
requests
websocket-client
python-dotenv